Revert "Added Galleries and Bylaws (#39)"
This reverts commit e594fc792abc008d3e1676ccd4ea32446af1e829.
