+++ package.json
@@ -19,13 +19,13 @@
     "fs": "^0.0.1-security",
     "googleapis": "^139.0.0",
     "gray-matter": "^4.0.3",
-    "mammoth": "^1.9.0",
     "nepali-date-converter": "^3.3.2",
     "net": "^1.0.2",
     "next": "^14.2.1",
     "next-mdx-remote": "^4.4.1",
     "react": "^18",
     "react-dom": "^18",
+    "sass": "^1.79.3",
     "react-firebase-hooks": "^5.1.1",
     "react-papaparse": "^4.4.0",
     "sass": "^1.79.3"
