--- src/app/Bylaws/page.jsx
@@ -1,13 +1,0 @@
-import Header from "../../components/Header/Header";
-import Bylaws from "../../components/Bylaws/Bylaws"
-import Footer from "../../components/Footer/Footer"
-
-export default function Ab() {
-  return (
-    <main class="flex flex-col h-dvh ">
-      <Header/>
-      <Bylaws/>
-      <Footer/>
-    </main>
-  );
-}
