--- src/app/Gallery2024/page.jsx
@@ -1,13 +1,0 @@
-import Header from "../../components/Header/Header";
-import Gallery from "../../components/Gallery/Gallery";
-import Footer from "../../components/Footer/Footer";
-
-export default function Ab() {
-  return (
-    <main class="flex flex-col ">
-      <Header/>
-      <Gallery/>
-      <Footer/>
-    </main>
-  );
-}
