--- src/app/GalleryYearSelection/page.jsx
@@ -1,13 +1,0 @@
-import Header from "../../components/Header/Header";
-import GalleryYearSelection from "../../components/Gallery/GalleryYearSelection";
-import Footer from "../../components/Footer/Footer";
-
-export default function Ab() {
-  return (
-    <main class="flex flex-col ">
-      <Header/>
-      <GalleryYearSelection/>
-      <Footer/>
-    </main>
-  );
-}
