+++ src/components/Body/Hero.jsx
@@ -35,13 +35,9 @@ const Hero = () => {
                                     <button class="bg-cyan-500 dark:bg-slate-800 shadow-gray-700 dark:text-white 
                             dark:shadow-slate-950   shadow-lg rounded-lg p-2 text-white m-2 drop-shadow-2xl">LINS News</button>
                                 </Link>
-                                <Link href="/GalleryYearSelection">
+                                <Link href="/Gallery">
                                     <button class="bg-cyan-500 dark:bg-slate-800 shadow-gray-700 dark:text-white 
                             dark:shadow-slate-950   shadow-lg rounded-lg p-2 text-white m-2 drop-shadow-2xl">Gallery</button>
-                                </Link>
-                                <Link href="/Bylaws">
-                                    <button class="bg-cyan-500 dark:bg-slate-800 shadow-gray-700 dark:text-white 
-                            dark:shadow-slate-950   shadow-lg rounded-lg p-2 text-white m-2 drop-shadow-2xl">Bylaws</button>
                                 </Link>
                             </div>
                         </div>
@@ -76,13 +72,9 @@ const Hero = () => {
                                     <button class="bg-cyan-500 dark:bg-slate-800 shadow-gray-700  dark:shadow-slate-950  
                              shadow-lg dark:text-white rounded-lg p-2 text-white m-2 drop-shadow-2xl">LINS News</button>
                                 </Link>
-                                <Link href="/GalleryYearSelection">
+                                <Link href="/Gallery">
                                     <button class="bg-cyan-500 dark:bg-slate-800 shadow-gray-700  dark:shadow-slate-950  
                              shadow-lg dark:text-white rounded-lg p-2 text-white m-2 drop-shadow-2xl">Gallery</button>
-                                </Link>
-                                <Link href="/Bylaws">
-                                    <button class="bg-cyan-500 dark:bg-slate-800 shadow-gray-700  dark:shadow-slate-950  
-                             shadow-lg dark:text-white rounded-lg p-2 text-white m-2 drop-shadow-2xl">Bylaws</button>
                                 </Link>
                             </div>
                         </div>
