--- src/components/Gallery/2024/BloodDrive.js
@@ -1,33 +1,0 @@
-const item = {
-    images: [
-        { src: "/images/2024/Blood Drive/1.jpg", alt: "Image 1" },
-        { src: "/images/2024/Blood Drive/2.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/3.jpg", alt: "Image 1" },
-        { src: "/images/2024/Blood Drive/4.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/5.jpg", alt: "Image 1" },
-        { src: "/images/2024/Blood Drive/6.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/7.jpg", alt: "Image 1" },
-        { src: "/images/2024/Blood Drive/8.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/9.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/10.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/11.jpg", alt: "Image 1" },
-        { src: "/images/2024/Blood Drive/12.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/13.jpg", alt: "Image 1" },
-        { src: "/images/2024/Blood Drive/14.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/15.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/16.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/17.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/18.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/19.jpg", alt: "Image 2" },
-        { src: "/images/2024/Blood Drive/20.jpg", alt: "Image 2" },
-
-    ],
-
-    description: {
-        englishProgramName: "Blood Drive Program",
-        nepaliProgramName: "रक्त दान कार्यक्रम"
-            }
-
-}
-
-export default item
