--- src/components/Gallery/2024/ImmigrationRightsInteractionProgram.js
@@ -1,28 +1,0 @@
-const item = {
-    images: [
-        { src: "/images/2024/Immigration Rights Interaction Program/1.jpg", alt: "Image 1" },
-        { src: "/images/2024/Immigration Rights Interaction Program/2.jpg", alt: "Image 2" },
-        { src: "/images/2024/Immigration Rights Interaction Program/3.jpg", alt: "Image 1" },
-        { src: "/images/2024/Immigration Rights Interaction Program/4.jpg", alt: "Image 2" },
-        { src: "/images/2024/Immigration Rights Interaction Program/5.jpg", alt: "Image 1" },
-        { src: "/images/2024/Immigration Rights Interaction Program/6.jpg", alt: "Image 2" },
-        { src: "/images/2024/Immigration Rights Interaction Program/7.jpg", alt: "Image 1" },
-        { src: "/images/2024/Immigration Rights Interaction Program/8.jpg", alt: "Image 2" },
-        { src: "/images/2024/Immigration Rights Interaction Program/9.jpg", alt: "Image 2" },
-        { src: "/images/2024/Immigration Rights Interaction Program/10.jpg", alt: "Image 2" },
-        { src: "/images/2024/Immigration Rights Interaction Program/11.jpg", alt: "Image 1" },
-        { src: "/images/2024/Immigration Rights Interaction Program/12.jpg", alt: "Image 2" },
-        { src: "/images/2024/Immigration Rights Interaction Program/13.jpg", alt: "Image 1" },
-        { src: "/images/2024/Immigration Rights Interaction Program/14.jpg", alt: "Image 2" },
-        { src: "/images/2024/Immigration Rights Interaction Program/15.jpg", alt: "Image 2" },
-
-    ],
-
-    description: {
-        englishProgramName: "Immigration - Rights Interaction Program",
-        nepaliProgramName: "आव्रजन - अधिकार अन्तर्क्रिया कार्यक्रम"
-            }
-
-}
-
-export default item
