--- src/components/Gallery/2024/NepaliNewYear2081.js
@@ -1,33 +1,0 @@
-const item = {
-    images: [
-        { src: "/images/2024/Nepali Dashin Program 2081/1.jpg", alt: "Image 1" },
-        { src: "/images/2024/Nepali Dashin Program 2081/20.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/19.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/2.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/3.jpg", alt: "Image 1" },
-        { src: "/images/2024/Nepali Dashin Program 2081/4.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/5.jpg", alt: "Image 1" },
-        { src: "/images/2024/Nepali Dashin Program 2081/6.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/7.jpg", alt: "Image 1" },
-        { src: "/images/2024/Nepali Dashin Program 2081/8.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/9.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/10.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/11.jpg", alt: "Image 1" },
-        { src: "/images/2024/Nepali Dashin Program 2081/12.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/12.jpg", alt: "Image 1" },
-        { src: "/images/2024/Nepali Dashin Program 2081/14.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/15.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/16.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/17.jpg", alt: "Image 2" },
-        { src: "/images/2024/Nepali Dashin Program 2081/18.jpg", alt: "Image 2" },
-
-    ],
-
-    description: {
-        englishProgramName: "Nepali Dashain Program 2081",
-        nepaliProgramName: "नेपाली दशैं कार्यक्रम २०८१"
-            }
-
-}
-
-export default item
